+++ src/OpenTelemetry.Exporter.Console/ConsoleExporterHelperExtensions.cs
@@ -39,7 +39,7 @@ namespace OpenTelemetry.Trace
 
             var options = new ConsoleExporterOptions();
             configure?.Invoke(options);
-            return builder.AddProcessor(new SimpleExportProcessor<Activity>(new ConsoleActivityExporter(options)));
+            return builder.AddProcessor(new SimpleActivityExportProcessor(new ConsoleActivityExporter(options)));
         }
     }
 }