+++ src/OpenTelemetry.Instrumentation.SqlClient/SqlClientInstrumentationOptions.cs
@@ -30,13 +30,35 @@ namespace OpenTelemetry.Instrumentation.SqlClient
     {
         /*
          * Match...
+         *  protocol[ ]:[ ]serverName
          *  serverName
-         *  serverName[ ]\\[ ]instanceName
+         *  serverName[ ]\[ ]instanceName
          *  serverName[ ],[ ]port
-         *  serverName[ ]\\[ ]instanceName[ ],[ ]port
+         *  serverName[ ]\[ ]instanceName[ ],[ ]port
+         *
          * [ ] can be any number of white-space, SQL allows it for some reason.
+         *
+         * Optional "protocol" can be "tcp", "lpc" (shared memory), or "np" (named pipes). See:
+         *  https://docs.microsoft.com/troubleshoot/sql/connect/use-server-name-parameter-connection-string, and
+         *  https://docs.microsoft.com/dotnet/api/system.data.sqlclient.sqlconnection.connectionstring?view=dotnet-plat-ext-5.0
+         *
+         * In case of named pipes the Data Source string can take form of:
+         *  np:serverName\instanceName, or
+         *  np:\\serverName\pipe\pipeName, or
+         *  np:\\serverName\pipe\MSSQL$instanceName\pipeName - in this case a separate regex (see NamedPipeRegex below)
+         *  is used to extract instanceName
          */
-        private static readonly Regex DataSourceRegex = new Regex("^(.*?)\\s*(?:[\\\\,]|$)\\s*(.*?)\\s*(?:,|$)\\s*(.*)$", RegexOptions.Compiled);
+        private static readonly Regex DataSourceRegex = new Regex("^(.*\\s*:\\s*\\\\{0,2})?(.*?)\\s*(?:[\\\\,]|$)\\s*(.*?)\\s*(?:,|$)\\s*(.*)$", RegexOptions.Compiled);
+
+        /// <summary>
+        /// In a Data Source string like "np:\\serverName\pipe\MSSQL$instanceName\pipeName" match the
+        /// "pipe\MSSQL$instanceName" segment to extract instanceName if it is available.
+        /// </summary>
+        /// <see>
+        /// <a href="https://docs.microsoft.com/previous-versions/sql/sql-server-2016/ms189307(v=sql.130)"/>
+        /// </see>
+        private static readonly Regex NamedPipeRegex = new Regex("pipe\\\\MSSQL\\$(.*?)\\\\", RegexOptions.Compiled);
+
         private static readonly ConcurrentDictionary<string, SqlConnectionDetails> ConnectionDetailCache = new ConcurrentDictionary<string, SqlConnectionDetails>(StringComparer.OrdinalIgnoreCase);
 
         // .NET Framework implementation uses SqlEventSource from which we can't reliably distinguish
@@ -58,17 +80,17 @@ namespace OpenTelemetry.Instrumentation.SqlClient
         /// When using <c>System.Data.SqlClient</c>, the instrumentation will only capture <c>sqlCommand.CommandText</c> for <see cref="CommandType.StoredProcedure"/> commands.
         /// </para>
         /// </remarks>
-        public bool SetStatementText { get; set; }
+        public bool SetDbStatement { get; set; }
 #else
         /// <summary>
         /// Gets or sets a value indicating whether or not the <see cref="SqlClientInstrumentation"/> should add the names of <see cref="CommandType.StoredProcedure"/> commands as the <see cref="SemanticConventions.AttributeDbStatement"/> tag. Default value: True.
         /// </summary>
-        public bool SetStoredProcedureCommandName { get; set; } = true;
+        public bool SetDbStatementForStoredProcedure { get; set; } = true;
 
         /// <summary>
         /// Gets or sets a value indicating whether or not the <see cref="SqlClientInstrumentation"/> should add the text of <see cref="CommandType.Text"/> commands as the <see cref="SemanticConventions.AttributeDbStatement"/> tag. Default value: False.
         /// </summary>
-        public bool SetTextCommandContent { get; set; }
+        public bool SetDbStatementForText { get; set; }
 #endif
 
         /// <summary>
@@ -86,7 +108,7 @@ namespace OpenTelemetry.Instrumentation.SqlClient
         /// <para><see cref="Activity"/>: the activity being enriched.</para>
         /// <para>string: the name of the event.</para>
         /// <para>object: the raw <c>SqlCommand</c> object from which additional information can be extracted to enrich the activity.</para>
-        /// <para>See also: <a href="https://github.com/open-telemetry/opentelemetry-dotnet/tree/master/src/OpenTelemetry.Instrumentation.SqlClient#Enrich">example</a>.</para>
+        /// <para>See also: <a href="https://github.com/open-telemetry/opentelemetry-dotnet/tree/main/src/OpenTelemetry.Instrumentation.SqlClient#Enrich">example</a>.</para>
         /// </remarks>
         /// <example>
         /// <code>
@@ -112,7 +134,7 @@ namespace OpenTelemetry.Instrumentation.SqlClient
         /// Gets or sets a value indicating whether the exception will be recorded as ActivityEvent or not. Default value: False.
         /// </summary>
         /// <remarks>
-        /// https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/semantic_conventions/exceptions.md.
+        /// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/exceptions.md.
         /// </remarks>
         public bool RecordException { get; set; }
 #endif
@@ -121,8 +143,10 @@ namespace OpenTelemetry.Instrumentation.SqlClient
         {
             Match match = DataSourceRegex.Match(dataSource);
 
-            string serverHostName = match.Groups[1].Value;
+            string serverHostName = match.Groups[2].Value;
             string serverIpAddress = null;
+
+            string instanceName;
 
             var uriHostNameType = Uri.CheckHostName(serverHostName);
             if (uriHostNameType == UriHostNameType.IPv4 || uriHostNameType == UriHostNameType.IPv6)
@@ -131,25 +155,56 @@ namespace OpenTelemetry.Instrumentation.SqlClient
                 serverHostName = null;
             }
 
-            string instanceName;
-            string port;
-            if (match.Groups[3].Length > 0)
+            string maybeProtocol = match.Groups[1].Value;
+            bool isNamedPipe = maybeProtocol.Length > 0 &&
+                               maybeProtocol.StartsWith("np", StringComparison.OrdinalIgnoreCase);
+
+            if (isNamedPipe)
             {
-                instanceName = match.Groups[2].Value;
-                port = match.Groups[3].Value;
+                string pipeName = match.Groups[3].Value;
+                if (pipeName.Length > 0)
+                {
+                    var namedInstancePipeMatch = NamedPipeRegex.Match(pipeName);
+                    if (namedInstancePipeMatch.Success)
+                    {
+                        instanceName = namedInstancePipeMatch.Groups[1].Value;
+                        return new SqlConnectionDetails
+                        {
+                            ServerHostName = serverHostName,
+                            ServerIpAddress = serverIpAddress,
+                            InstanceName = instanceName,
+                            Port = null,
+                        };
+                    }
+                }
+
+                return new SqlConnectionDetails
+                {
+                    ServerHostName = serverHostName,
+                    ServerIpAddress = serverIpAddress,
+                    InstanceName = null,
+                    Port = null,
+                };
+            }
+
+            string port;
+            if (match.Groups[4].Length > 0)
+            {
+                instanceName = match.Groups[3].Value;
+                port = match.Groups[4].Value;
                 if (port == "1433")
                 {
                     port = null;
                 }
             }
-            else if (int.TryParse(match.Groups[2].Value, out int parsedPort))
+            else if (int.TryParse(match.Groups[3].Value, out int parsedPort))
             {
-                port = parsedPort == 1433 ? null : match.Groups[2].Value;
+                port = parsedPort == 1433 ? null : match.Groups[3].Value;
                 instanceName = null;
             }
             else
             {
-                instanceName = match.Groups[2].Value;
+                instanceName = match.Groups[3].Value;
 
                 if (string.IsNullOrEmpty(instanceName))
                 {