+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -6,13 +6,15 @@
   on .NET Framework.
   ([#1599](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1599))
 * SqlClientInstrumentationOptions API changes: `SetStoredProcedureCommandName`
-  and `SetTextCommandContent` are now only available on .NET Core. On .NET
-  Framework they are replaced by a single `SetStatementText` property.
+  and `SetTextCommandContent` have been renamed to
+  `SetDbStatementForStoredProcedure` and `SetDbStatementForText`. They are now
+  only available on .NET Core. On .NET Framework they are replaced by a single
+  `SetDbStatement` property.
 * On .NET Framework, "db.statement_type" attribute is no longer set for
   activities created by the instrumentation.
 * New setting on SqlClientInstrumentationOptions on .NET Core: `RecordException`
   can be set to instruct the instrumentation to record SqlExceptions as Activity
-  [events](https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/semantic_conventions/exceptions.md).
+  [events](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/exceptions.md).
   ([#1592](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1592))
 
 ## 1.0.0-rc1.1